[RENAME]/#4 UserController로 수정
